python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmx"
version = "0.1.0"
description = "Python client for the CMX orchestration daemon"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
